fix: API 공용 타입 shared로 이동
