+++ src/shared/consts/index.ts
@@ -1,1 +1,2 @@
 export { QUERY_KEYS } from './query-key/queryKeys';
+export type * from './types';
