+++ src/shared/consts/types/comment.d.ts
@@ -1,0 +1,9 @@
+export interface Comment {
+  id: number;
+
+  user: User;
+  content: string;
+
+  createdAt: string;
+  updatedAt: string;
+}
