+++ src/shared/consts/types/feed.d.ts
@@ -1,0 +1,23 @@
+export interface Feed {
+  id: number;
+
+  user: User;
+
+  title: string;
+  content: string;
+  images: Image[];
+
+  likeCount: number;
+  commentCount: number;
+
+  isLiked: boolean;
+  isBookmark: boolean;
+
+  createdAt: string;
+  updatedAt: string;
+}
+
+export interface Image {
+  id: number;
+  imageUrl: string;
+}
