+++ src/shared/consts/types/index.ts
@@ -1,0 +1,4 @@
+export type { Feed, Image } from './feed';
+export type { Comment } from './comment';
+export type { Like } from './like';
+export type { User } from './user';
