+++ src/shared/consts/types/like.d.ts
@@ -1,0 +1,4 @@
+export interface Like {
+  totalCount: number;
+  isLiked: boolean;
+}
