+++ src/shared/consts/types/user.d.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: number;
+  profileImage: string;
+  name: string;
+  content: string;
+}
